requirements.txt
python-etcd
IPy
mock
python-novaclient
PyYAML
netifaces
consulate
tornado
python-hpilo
